requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.10.0
black==23.12.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
exceptiongroup==1.3.0
fastapi==0.104.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
mypy_extensions==1.1.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pypyr==5.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.3.0
tomli_w==1.2.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
